Add buttons
It contains a parent <View> with three buttons aligned in a row. The
button in the middle with the plus icon (+) will open the modal and is
styled differently than the other two buttons.

1. Inside the components directory, create a new CircleButton.tsx file

To render the plus icon, this button uses the <MaterialIcons> icon set
from the @expo/vector-icons library.

The other two buttons also use <MaterialIcons> to display vertically
aligned text labels and icons. Create a file named IconButton.tsx inside
the components directory. This component accepts three props:

- icon: the name corresponding to the MaterialIcons library icon.
- label: the text label displayed on the button.
- onPress: this function invokes when the user presses the button.

Inside app/(tabs)/index.tsx:

Import the CircleButton and IconButton components to display them. Add
three placeholder functions for these buttons. The onReset() function
invokes when the user presses the reset button, causing the image picker
button to appear again. We'll add the functionality for the other two
functions later.

Now pressing "Use this photo" will display the CircleButton component
flanked by a Reset button on the left and a Save button on the right.