Display a list of emoji
Let's add a horizontal list of emoji in the modal's content. We'll use
the <FlatList> component from React Native for it.

Create a EmojiList.tsx file inside the components directory

Let's learn what the EmojiList code does:

- The <FlatList> component above renders all the emoji images using the
  Image component, wrapped by a <Pressable>. Later, we will improve it
  so that the user can tap an emoji on the screen to make it appear as a
  sticker on the image.

- It also takes an array of items provided by the emoji array variable
  as the value of the data prop. The renderItem prop takes the item from
  the data and returns the item in the list. Finally, we added Image and
  the <Pressable> components to display this item.

- The horizontal prop renders the list horizontally instead of
  vertically. The showsHorizontalScrollIndicator uses React Native's
  Platform module to check the value and display the horizontal scroll
  bar on web.

Update the app/(tabs)/index.tsx to import the <EmojiList> component and
replace the comments inside the <EmojiPicker> component

In the EmojiList component, the onSelect prop selects the emoji and
after selecting it, the onCloseModal closes the modal.