Display the selected emoji
Now, we'll put the emoji sticker on the image. Create a new file in the
components directory and call it EmojiSticker.tsx.

This component receives two props:

- imageSize: a value defined inside the Index component. We will use
  this value in the next chapter to scale the image's size when tapped.

- stickerSource: the source of the selected emoji image.

Import this component in the app/(tabs)/index.tsx file and update the
Index component to display the emoji sticker on the image. We'll check
if the pickedEmoji state is not undefined and, if not, add a sticker to
the view.