1. Add GestureHandlerRootView
To get gesture interactions to work in the app, we'll render
<GestureHandlerRootView> from react-native-gesture-handler at the top of
Index component. Replace the root level <View> component in the
app/(tabs)/index.tsx with <GestureHandlerRootView>.

2. Use animated components

An Animated component looks at the style prop of the component and
determines which values to animate and apply updates to create an
animation. Reanimated exports animated components such as
<Animated.View>, <Animated.Text>, or <Animated.ScrollView>. We will
apply animations to the <Animated.Image> component to make a double tap
gesture work.

Open the EmojiSticker.tsx file in the components directory. Inside it,
import Animated from the react-native-reanimated library to use animated
components.

Replace the Image component with <Animated.Image>.