Add a tap gesture
React Native Gesture Handler allows us to add behavior when it detects
touch input, like a double tap event.

In the EmojiSticker.tsx file:

Import Gesture and GestureDetector from react-native-gesture-handler.

To recognize the tap on the sticker, import useAnimatedStyle,
useSharedValue, and withSpring from react-native-reanimated to animate
the style of the <Animated.Image>.

Inside the EmojiSticker component, create a reference called scaleImage
using the useSharedValue() hook. It will take the value of imageSize as
its initial value.

Creating a shared value using the useSharedValue() hook has many
advantages. It helps to mutate data and runs animations based on the
current value. We can access and modify the shared value using the
.value property.