Add a tap gesture
To animate the transition, let's use a spring-based animation. This will
make it feel alive because it's based on the real-world physics of a
spring. We will use the withSpring() function provided by
react-native-reanimated.

On the sticker image, we'll use the useAnimatedStyle() hook to create a
style object. This will help us to update styles using shared values
when the animation happens. We'll also scale the size of the image by
manipulating the width and height properties. The initial values of
these properties are set to imageSize.