Add a pan gesture
To recognize a dragging gesture on the sticker and to track its
movement, we'll use a pan gesture. In the components/EmojiSticker.tsx:

- Create two new shared values: translateX and translateY.

- Replace the <View> with the <Animated.View> component.

The translation values defined will move the sticker around the screen.
Since the sticker moves along both axes, we need to track the X and Y
values.

In the useSharedValue() hooks, we have set both translation variables to
have an initial position of 0. This is the sticker's initial position
and a starting point. This value sets the sticker's initial position
when the gesture starts.

In the previous step, we triggered the onStart() callback for the tap
gesture chained to the Gesture.Tap() method. For the pan gesture,
specify an onChange() callback, which runs when the gesture is active
and moving.