Prompt for permissions
An app that requires sensitive information, such as accessing a device's
media library, has to prompt permission to allow or deny access. Using
usePermissions() hook from expo-media-library, we can use the permission
status and requestPermission() method to ask for access.

When the app loads for the first time and the permission status is
neither granted nor denied, the value of the status is null. When asked
for permission, a user can either grant the permission or deny it. We
can add a condition to check if it is null, and if it is, trigger the
requestPermission() method. After getting the access, the value of the
status changes to granted.

Note these must be installed:
`npx expo install react-native-view-shot expo-media-library`