Create a ref to save the current view
We'll use react-native-view-shot to allow the user to take a screenshot
within the app. This library captures the screenshot of a <View> as an
image using the captureRef() method. It returns the URI of the captured
screenshot image file.

1. Import captureRef from react-native-view-shot and useRef from React.

2. Create an imageRef reference variable to store the reference of the
screenshot image captured.

3. Wrap the <ImageViewer> and <EmojiSticker> components inside a <View>
and then pass the reference variable to it.

the collapsable prop is set to false. This allows the <View> component
to screenshot only of the background image and emoji sticker.