Capture a screenshot and save it
We can capture a screenshot of the view by calling the captureRef()
method from react-native-view-shot inside the onSaveImageAsync()
function. It accepts an optional argument where we can pass the width
and height of the screenshot capturing area. We can read more about
available options in the library's documentation.

The captureRef() method also returns a promise that fulfills with the
screenshot's URI. We will pass this URI as a parameter to
MediaLibrary.saveToLibraryAsync() and save the screenshot to the
device's media library.

Inside app/(tabs)/index.tsx, update the onSaveImageAsync() function