Initialize a development build
Install EAS CLI

We need to install the EAS Command Line Interface (CLI) tool as a global
dependency on our local machine. Run the following command:

    npm install -g eas-cli

Log in or sign up for an Expo account

If you have an Expo account and are signed in using Expo CLI, skip this
step. If you don't have an Expo account, sign up here and proceed with
the login command described below.

To log in, run the following command:

    eas login

This command asks for our Expo account email or username and password to
complete the login.

Initialize and link the project to EAS

For any new project, the first step is to initialize and link it to the
EAS servers. Run the following command:

    eas init

On running, this command:

Requests verification of the account owner by entering our Expo account
credentials and asks if we want to create a new EAS project:

Terminal

    ✔ Which account should own this project? > your-username
    ✔ Would you like to create a project for @your-username/sticker-smash? … yes
    ✔ Created @your-username/sticker-smash
    ✔ Project successfully linked (ID: XXXX-XX-XX-XXXX) (modified app.json)

Creates EAS project and provides a link to that project which we can
open in the Expo dashboard:

* New project in Expo dashboard

* Generates a unique projectId and links this EAS project to the example
app on our development machine.

* Modifies app.json to include extra.eas.projectId and updates its value
with the unique ID created.

What is projectId in app.json?

When eas init runs, it associates a unique identifier for our project in
app.json under extra.eas.projectId. The value of this property is used
to identify our project on EAS servers.

    {
      "extra": {
        "eas": {
          "projectId": "0cd3da2d-xxx-xxx-xxx-xxxxxxxxxx"
        }
      }
    }