Configure project for EAS Build
To set up our project for EAS Build, run the following command:

  eas build:configure

On running, this command:

* Prompts to select a platform: Android, iOS, or All. Since we are
creating Android and iOS apps, let's select All.

* Creates eas.json in the root of our project's directory with the
following configuration:

eas.json

  {
    "cli": {
      "version": ">= 14.2.0",
      "appVersionSource": "remote"
    },
    "build": {
      "development": {
        "developmentClient": true,
        "distribution": "internal"
      },
      "preview": {
        "distribution": "internal"
      },
      "production": {
        "autoIncrement": true
      }
    },
    "submit": {
      "production": {}
    }
  }

This is the default configuration for eas.json in a new project. It does
two things:

* Defines the current EAS CLI version.

* Adds three build profiles: development, preview, and production.

Explore development profile further

eas.json is a collection of different build profiles. Each profile is
tailored with distinct configurations to produce specific build types.
These profiles can also include platform-specific settings for Android
or iOS.

Currently, our focus is on the development profile, which includes the
following configuration:

* developmentClient: Enabled (true) for creating a debug build. It loads
the app using the expo-dev-client library, which provides development
tools and generates a build artifact for device or emulator/simulator
installation and allows using the app for local development as it
supports updating the JavaScript on the fly.

* distribution: Configured as internal to indicate that we want to share
the build internally (instead of uploading it on app stores).

Note: Builds offer extensive customization options, including
platform-specific settings and the ability to extend configurations
across different build profiles.