Create a build for the development profile
For Android, the development build must be in the .apk. While the
default Android format is .aab, which is ideal for Google Play Store
distribution, it cannot be installed on devices or emulators.

To create a .apk:

In eas.json, make sure that developmentClient is set to true under
build.development profile.

Then, run the eas build command with android as the platform and
development as the build profile:

  eas build --platform android --profile development

Tip: Next time you run eas build command, you can also use -p to specify
the platform. It is short for --platform.

This command prompts us with the following questions:

  What would you like your Android application id to be? Press return to
  select the default value provided for this prompt. This will add
  android.package in app.json.

  Generate a new Android Keystore? Press Y.

After responding, the build will queue up, and we can track its progress
via a provided link by the EAS CLI in the Expo dashboard:

Android development build details and progress in Expo dashboard

What information does a build details page contain?

The build details page displays the build type, profile, Expo SDK
version, app version, version code, last commit hash, and the identity
of the developer or account owner who initiated the build.

In the above image, the current status of the Build artifact shows that
the build is in progress. Upon completion, this section will offer an
option to download the build. The Logs outlines every step taken during
the Android build process on EAS Build. For the sake of brevity, we
won't explore each step in detail here. To learn more, see Android build
process.

What is an Android application ID?

Also known as the package name of our Android app, it stores the value
in DNS reverse notation format (com.owner.appname). Each component of
this notation should start with a lowercase letter.

For example, our example app has com.owner.stickersmash where com.owner
is the domain and stickersmash is our app name.