Create development build for iOS Simulator
Run the eas build command with ios as a platform and ios-simulator as
the build profile:

  eas build --platform ios --profile ios-simulator

This command prompts us with the following questions when we create the
build for the first time:

* What would you like your iOS bundle identifier to be? Press return to
select the default value provided for this prompt. This will add
ios.bundleIdentifier in app.json.

After responding to the prompts, our EAS Build is queued, and the EAS
CLI provides a link to view build details and track progress on the Expo
dashboard:

WHAT DOES A BUILD DETAILS PAGE CONTAIN?

The build details page displays the build type, profile, Expo SDK
version, app version, build number, last commit hash, and the identity
of the developer or account owner who initiated the build.

In the above image, the current status of the Build artifact shows that
the build is in progress. Upon completion, this section will offer an
option to download the build. The Logs outlines every step taken during
the iOS build process on EAS Build. For the sake of brevity, we won't
explore each step in detail here. To learn more, see iOS build process.

WHAT IS IOS BUNDLE IDENTIFIER?

The ios.bundleIdentifier is a unique name of our app. If we publish our
app right now, the Apple App Store will use this property and its value
to identify our app on the store.

This notation is defined as host.owner.app-name. For example, our
example app has com.owner.stickersmash where com.owner is the domain and
stickersmash is our app name.

INSTALL

In the terminal, once the build finishes, EAS CLI prompts us by asking
whether we want to run the build on an iOS Simulator. Press Y.

EAS CLI automatically gives option to run a build on iOS Simulator
Alternate: Use Expo Orbit You can use Expo Orbit to install the
development build. From Build artifact on the Expo dashboard, click Open
with Expo Orbit to install the development build on the iOS Simulator.

RUN

Start the development server by running the npx expo start command from
the project directory:

  npx expo start

Press i in the terminal window to open the project on the iOS Simulator.