Configure multiple app variants
RUN DEVELOPMENT SERVER

After builds are complete, follow the same procedure from previous
chapters to install them on a device or emulator/simulator.

Since we're identifying our development build with the APP_VARIANT
environment variable, we need to pass it to the command when starting
the development server. To do this, add a dev script in the "scripts"
field of our project's package.json:

    {
      "scripts": {
        "dev": "APP_VARIANT=development npx expo start"
      }
    }

Then Run the npm run dev command to start the development server:

    npm run dev

This script will evaluate app.config.js locally and load the environment
variable for the development profile.

Now, our development build will run on both Android and iOS, displaying
the modified app name from app.config.js. For example, the below
development build is running on an iOS Simulator. See that the app name
is StickerSmash (Dev):