Share previews with your team
Learn how to use EAS Update to send OTA updates and share previews with
a team.

Updates generally fix small bugs and push small changes in between app
store releases. They allow updating the non-native parts of our example
app, such as JavaScript code, styling, and images.

In this chapter, we'll use EAS Update to share changes with our team.
This will help us and our team quickly share previews of the change.

1. INSTALL EXPO-UPDATES LIBRARY

To initialize our project and send an update, we need to use the
expo-updates library. Run the following command to install it:

    npx expo install expo-updates