Share previews with your team
2. CONFIGURE EAS UPDATE

To initialize our project with EAS Update, we need to follow these
steps:

* Since we are using dynamic app.config.js for our app's configuration,
adding updates and runtimeVersion properties are required to make our
project compatible with EAS Update. Run the following command to
obtain these properties and their values from EAS and manually copy
them to app.config.js:

    eas update:configure

What about non-dynamic (app.json) projects?

If a project doesn't use dynamic app config (uses app.json instead of
app.config.js), the above command will configure our app to be
compatible with EAS Update and add the right properties to app.json and
eas.json.

Re-run eas update:configure to continue with the setup process. A
channel should be added to every build profile in eas.json:

    {
      "build": {
        "development": {
          "channel": "development"
        },
        "ios-simulator": {
        },
        "preview": {
          "channel": "preview"
        },
        "production": {
          "channel": "production"
        }
      }
    }

Notice that the eas update:configure command adds the channel to every
build profile in eas.json. However, our ios-simulator profile extends
the development profile and having a separate channel doesn't make
sense. We can safely remove ios-simulator.channel from the above
configuration.

What is a channel?

Channels are used to group builds together. If we have an Android and
iOS build, both on the app store, we can give them both a channel of
production. Later, we can tell EAS Update to target the production
channel, so our update will affect all builds with a production channel.

3. CREATE A DEVELOPMENT BUILD

We need to create a new development build since our last build doesn't
contain the expo-updates library. Run the following command:

    eas build --platform android --profile development

We are using a development build for Android devices to demonstrate
updates. However, we can use --platform all or --platform ios to create
a build for both platforms or just for iOS.

After the new version of the development build is created, make sure to
install it on a device.

4. MODIFY THE JAVASCRIPT CODE OF THE APP

Let's modify our example app's JavaScript code. If you are not using
Sticker Smash app, you can modify any piece of your code to see the
changes in the app.

We'll modify the text of the first button in our example app that says
Choose a photo to Select a photo.

App.js

    <Button theme="primary" label="Select a photo" onPress={pickImageAsync} />

5. PUBLISH AN UPDATE

Instead of creating a new build to share this change with our team for
testing, let's publish an update:

    eas update --branch development --message "Change first button label"

In the command above, we used the development branch. Every update is
associated with an update branch. It is similar to every commit that we
make with git, which is associated with a git branch.

By default, EAS will map branches and channels with the same name, if no
other mapping has been specified. So, by using the channel development
in our build profile and then publishing an update on the development
branch, we're asking EAS to deliver this update to builds with the
development channel. When we make an EAS Update branch, we can map it to
a channel.

After the update is published, the CLI will prompt us with information
about it.

Click on the Website link to see the Update on the Expo dashboard under
Updates

6. PREVIEW THE UPDATE LIVE IN A DEVELOPMENT BUILD

To preview the live update in a development build:

* Log in to your Expo account within the development build.
* Open the Extensions tab.
* Look for Branch: development listed under EAS Update.
* Tap on Open to access the update.

7. SHARING CHANGES WITH PREVIEW OR PRODUCTION BUILDS

Updates for non-development builds (preview or production) are
automatically downloaded to the device when the app starts up and makes
a request for any new updates.

Any team member running the preview or production build will receive the
update with the changes we push to those specific branches.

For example, for a preview build, we can run:

    eas update --branch preview --message "Change first button label"

Here is an example where we've published an update for the preview
build. To test the update, force close and reopen the app twice to
download and view the changes: