Trigger builds from a GitHub repository
1. CONFIGURE EXPO GITHUB APP

To use this functionality, we need to connect our GitHub account:

* In the Expo dashboard, go to expo.dev/settings, and under Connections
> GitHub, click Connect. This opens the Connect GitHub accounts page.

* Click the Get started button which opens a popup to authorize the Expo
GitHub app. Click Install and Authorize.

* Once the app is installed on our GitHub account, we need to link it to
our Expo account. In the next popup, click Link installation.

* Once the account is linked, it will show under GitHub.

2. CONNECT THE GITHUB REPOSITORY

To enable triggering builds from a GitHub repository, we need to connect
it to our project in the Expo dashboard:

* In Expo dashboard, go to Project > Configuration > GitHub.

* Under Connect a GitHub repository, we'll see a list of our GitHub
repos. We need to connect the right one. In the example, we're
searching for our repo sticker-smash.

* Click Connect for the project repository.

3. USE DEFAULT REPOSITORY SETTINGS

The Expo GitHub app needs to know where to find the source code of our
project. By default, it selects the root directory using /. In our
example project, the source code is also available in the root
repository. We can leave this to default in the Expo dashboard.

4. TRIGGER A BUILD USING A GITHUB PR LABEL

The Expo GitHub app provides us multiple options to trigger a build,
such as:

* Manually from the Builds page for a specific platform

* Automatically when new code is pushed to the repository

* Automatically using GitHub PR labels

To automatically trigger a build using a GitHub PR label, we're going to
utilize the third option from the list above:

* We need to specify the build image that we will be using. Open
eas.json, and under the development profile, add android.image and
ios.image properties and set their value to latest.

eas.json

    {
      "build": {
        "development": {
          ...
          "android": {
            "image": "latest"
          },
          "ios": {
            "image": "latest"
          }
        }
      }
      ...
    }

* Next, let's create a new branch called dev, and make a change in our
app's JavaScript code. Then, commit the change, push the branch, and
create a PR from that branch.

* In the PR link, under Labels, create a label called eas-build-all:development

* Click Create pull request button to create the PR. The Expo GitHub app
will start the process of creating a development build.

* In the Expo dashboard, on the Builds page, we can verify that the
builds for both Android and iOS are triggered.

If we check the details of an individual build, we can see under Created
by that the build is created by the GitHub app.